+++ activesupport/lib/active_support/string_inquirer.rb
@@ -17,7 +17,7 @@ module ActiveSupport
 
       def method_missing(method_name, *arguments)
         if method_name[-1] == '?'
-          self == method_name[0..-2]
+          self.casecmp(method_name[0..-2]) == 0
         else
           super
         end