+++ activesupport/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   `ActiveSupport::StringInquirer` is now case insensitive
+
+    *Ignatius Reza Lesmana*
+
 *   `ActiveSupport::Duration` supports weeks and hours.
 
         [1.hour.inspect, 1.hour.value, 1.hour.parts]